Moved sources-explorer into sources package
