+++ packages/bvaughn-architecture-demo/components/sources/SourceExplorer.module.css
