+++ packages/bvaughn-architecture-demo/components/sources/SourceExplorer.tsx
